backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
python-dotenv==1.0.1
langchain==0.3.7
langchain-google-genai==2.0.2
langchain-community==0.3.3
google-generativeai==0.8.5
yfinance==0.2.66
duckduckgo-search==7.4.0
langchain-experimental==0.3.4
requests==2.32.3